capa/main.py: add a `capabilities` module and move all of the capability extraction there
